chore: change feature flag fetch method
