+++ frontend/src/scenes/billing/BillingProductAddon.tsx
@@ -7,7 +7,7 @@ import { LemonSelectOptions, LemonTag, Link, Tooltip } from '@posthog/lemon-ui'
 
 import { TRIAL_CANCELLATION_SURVEY_ID, UNSUBSCRIBE_SURVEY_ID } from 'lib/constants'
 import { humanFriendlyCurrency } from 'lib/utils'
-import { getProductIcon } from 'scenes/onboarding/productSelection/ProductSelection'
+import { getProductIcon } from 'scenes/onboarding/utils'
 
 import { BillingProductV2AddonType } from '~/types'
